Create a Point of Sales (POS) system for the Super-Saving supermarket chain.
The cashier enters the item code and amounts to add the grocery item to the
bill. Given the item code, the price, weight/size of the product, date of manufacturing and expiry,
and manufacturer name are fetched from the database. In this lab, your database is a CSV file given
with the lab description. In addition, from time to time, each item is given a discount which varies
from 0-75%. The system processes the entered item list and produces the bill. The bill contains the
cashier’s name, branch, customer name (if a registered customer), item list (unit price, quantity,
discount, net price), total discount, total cost, date, and time. The date and time should indicate
when the bill was printed. The bill should be saved as a PDF file to be printed for the customer.
Further, this POS system is required to handle pending bills. For example, suppose a customer
brings a cart full of groceries to the cashier but forgets to weigh some vegetables and fruits. He
realizes this only after the cashier has entered half of the items into the system. While this customer
goes and weighs the vegetables and fruits he purchases, the cashier wants to keep his bill as a
pending one and deal with the billing for other customers. Furthermore, the management requires
the system to analyze bills and generate a total revenue report for a given date range. The generated
report should be emailed to the supermarket sales team at salesteam@supersaving.lk.